refactor(project): :truck: reorganise components
